+++ pykilosort/__init__.py
@@ -10,9 +10,10 @@ else:
 
 from .utils import Bunch, memmap_binary_file, read_data, load_probe, plot_dissimilarity_matrices, plot_diagnostics
 from .main import run, run_export, run_spikesort, run_preprocess
+from .io.probes import np1_probe, np2_probe, np2_4shank_probe
 
 
-__version__ = 'ibl_1.1.0'
+__version__ = 'ibl_1.2.0'
 
 # Set a null handler on the root logger
 logger = logging.getLogger(__name__)