+++ pykilosort/learn.py
@@ -1,4 +1,5 @@
 import logging
+import os
 from math import sqrt, ceil
 
 import numpy as np
@@ -196,6 +197,25 @@ def getMeWtW2(W, U0, Nnearest=None):
         return WtW
 
 
+def custom_lexsort(arrays):
+    """
+    Lexsort of 1D Cupy arrays, last array given is used for the primary sort order, second-last
+    for secondary sort order and so on
+    :param arrays: List of 1D Cupy arrays, all lengths must match
+    :return: Cupy array of indices that sort the arrays according to the above
+    """
+    # Check arrays are one-dimensional and have equal length
+    for array in arrays:
+        assert array.ndim == 1
+    n = arrays[0].shape[0]
+    for array in arrays[1:]:
+        assert array.shape[0] == n
+
+    # Concatenate arrays and pass to cupy lexsort function
+    lex_array = cp.concatenate([array.reshape(1, -1) for array in arrays], axis=0)
+    return cp.lexsort(lex_array)
+
+
 def mexGetSpikes2(Params, drez, wTEMP, iC):
     code, constants = get_cuda("mexGetSpikes2")
 
@@ -265,6 +285,12 @@ def mexGetSpikes2(Params, drez, wTEMP, iC):
         (32,),
         (d_Params, d_x, d_st, d_id, d_st1, d_id1, d_counter),
     )
+
+    # Order of peaks found can vary depending on thread execution times, sort for determinism
+    n_spikes = int(min(maxFR, d_counter[1]))
+    sorted_ids = custom_lexsort([d_id1[:n_spikes], d_st1[:n_spikes]])
+    d_st1[:n_spikes] = d_st1[:n_spikes][sorted_ids]
+    d_id1[:n_spikes] = d_id1[:n_spikes][sorted_ids]
 
     # add new spikes to 2nd counter
     counter[0] = d_counter[1]
@@ -421,7 +447,7 @@ def mexMPnu8(Params, dataRAW, U, W, mu, iC, iW, UtU, iList, wPCA, params):
         (d_Params, d_dout, d_mu, d_err, d_eloss, d_ftype),
     )
 
-    if params.stablemode_enabled and not params.deterministicmode_enabled:
+    if params.stable_mode and not params.deterministic_mode:
         d_draw64 = cp.array(d_draw, dtype=np.float64)
 
     # loop to find and subtract spikes
@@ -463,8 +489,8 @@ def mexMPnu8(Params, dataRAW, U, W, mu, iC, iW, UtU, iList, wPCA, params):
                 (d_Params, d_st, d_id, d_counter, d_dout, d_iList, d_mu, d_feat),
             )
 
-        if params.deterministicmode_enabled:
-            if params.stablemode_enabled:
+        if params.deterministic_mode:
+            if params.stable_mode:
                 d_stSort = d_st[counter[1]:counter[0]] # cudaMemcpy( d_stSort, d_st+counter[1], (counter[0] - counter[1])*sizeof(int), cudaMemcpyDeviceToDevice );
                 d_idx[:counter[0]-counter[1]] = cp.argsort(d_stSort) # cdp_simple_quicksort<<< 1, 1 >>>(d_stSort, d_idx, 0, counter[0] - counter[1] - 1, 0);
             else:
@@ -506,7 +532,7 @@ def mexMPnu8(Params, dataRAW, U, W, mu, iC, iW, UtU, iList, wPCA, params):
                 ),
             )
         else:
-            if params.stablemode_enabled:
+            if params.stable_mode:
                 subtract_spikes_v4 = cp.RawKernel(code, "subtract_spikes_v4")
                 subtract_spikes_v4(
                     (Nfilt,),
@@ -590,7 +616,7 @@ def mexMPnu8(Params, dataRAW, U, W, mu, iC, iW, UtU, iList, wPCA, params):
         # update 1st counter from 2nd counter
         d_counter[1] = d_counter[0]
 
-    if params.stablemode_enabled and not params.deterministicmode_enabled:
+    if params.stable_mode and not params.deterministic_mode:
         d_draw = cp.array(d_draw64, dtype=np.float32)
 
     # compute PC features from residuals + subtractions
@@ -617,7 +643,7 @@ def mexMPnu8(Params, dataRAW, U, W, mu, iC, iW, UtU, iList, wPCA, params):
             ),
         )
 
-    if params.stablemode_enabled:
+    if params.stable_mode:
         # d_idx = array of time sorted indices
         d_idx[:counter[0]] = cp.argsort(d_st[:counter[0]]) # cdp_simple_quicksort<<< 1, 1 >>>(d_stSort, d_idx, 0, counter[0] - counter[1] - 1, 0);
     else:
@@ -653,16 +679,23 @@ def mexMPnu8(Params, dataRAW, U, W, mu, iC, iW, UtU, iList, wPCA, params):
 
     del d_counter, d_Params, d_ftype, d_err, d_eloss, d_z, d_dout, d_data
 
+    # Sort to ensure determinism as different thread execution times leads to spikes being found
+    # in a random order
+
+    sorted_idx = custom_lexsort([d_y[:minSize], d_id[:minSize], d_st[:minSize]])
+
+    # Indexing a f-contiguous array returns a c-contiguous array so we need to recast back into
+    # fortran order
     return (
-        d_st[:minSize],
-        d_id[:minSize],
-        d_y[:minSize],
-        d_feat[..., :minSize],
+        d_st[:minSize][sorted_idx],
+        d_id[:minSize][sorted_idx],
+        d_y[:minSize][sorted_idx],
+        cp.asfortranarray(d_feat[..., :minSize][..., sorted_idx]),
         d_dWU,
         d_draw,
         d_nsp,
-        d_featPC[..., :minSize],
-        d_x[:minSize],
+        cp.asfortranarray(d_featPC[..., :minSize][..., sorted_idx]),
+        d_x[:minSize][sorted_idx],
     )
 
 
@@ -755,9 +788,9 @@ def triageTemplates2(params, iW, C2C, W, U, dWU, mu, nsp, ndrop):
     idrop = nsp < m0  # drop any templates with firing rate below this
 
     # remove those templates everywhere
-    W = W[:, ~idrop, :]
-    U = U[:, ~idrop, :]
-    dWU = dWU[:, :, ~idrop]
+    W = cp.asfortranarray(W[:, ~idrop, :])
+    U = cp.asfortranarray(U[:, ~idrop, :])
+    dWU = cp.asfortranarray(dWU[:, :, ~idrop])
     mu = mu[~idrop]
     nsp = nsp[~idrop]
     # keep track of how many templates have been removed this way
@@ -782,9 +815,9 @@ def triageTemplates2(params, iW, C2C, W, U, dWU, mu, nsp, ndrop):
     idrop = amax > 0
 
     # remove these templates everywhere like before
-    W = W[:, ~idrop, :]
-    U = U[:, ~idrop, :]
-    dWU = dWU[:, :, ~idrop]
+    W = cp.asfortranarray(W[:, ~idrop, :])
+    U = cp.asfortranarray(U[:, ~idrop, :])
+    dWU = cp.asfortranarray(dWU[:, :, ~idrop])
     mu = mu[~idrop]
     nsp = nsp[~idrop]
     # keep track of how many templates have been removed this way
@@ -837,14 +870,8 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
     # decay of gaussian spatial mask centered on a channel
     sigmaMask = params.sigmaMask
 
-    batchstart = list(range(0, NT * nBatches + 1, NT))
-
     # find the closest NchanNear channels, and the masks for those channels
     iC, mask, C2C = getClosestChannels(probe, sigmaMask, NchanNear)
-
-    # sorting order for the batches
-    isortbatches = iorig
-    nhalf = int(ceil(nBatches / 2)) - 1  # halfway point
 
     # batch order schedule is a random permutation of all batches
     ischedule = np.random.permutation(nBatches)
@@ -935,18 +962,15 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
         ctx.path("fWpc", ext=".dat"), dtype=np.float32, shape=(NchanNear, Nrank, -1)
     )
 
+    if params.low_memory:
+        feature_path = ctx.context_path / 'spike_features'
+        os.mkdir(feature_path)
+
     for ibatch in tqdm(range(niter), desc="Optimizing templates"):
         # korder is the index of the batch at this point in the schedule
         korder = int(irounds[ibatch])
         # k is the index of the batch in absolute terms
-        k = int(isortbatches[korder])
         logger.debug("Batch %d/%d, %d templates.", ibatch, niter, Nfilt)
-
-        if ibatch > niter - nBatches - 1 and korder == nhalf:
-            # this is required to revert back to the template states in the middle of the
-            # batches
-            W, dWU = ir.W, ir.dWU
-            logger.debug("Reverted back to middle timepoint.")
 
         if ibatch < niter - nBatches:
             # obtained pm for this batch
@@ -954,7 +978,7 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
             pm = pmi[ibatch] * ones((Nfilt,), dtype=np.float64, order="F")
 
         # loading a single batch
-        dataRAW = ir.data_loader.load_batch(k)
+        dataRAW = ir.data_loader.load_batch(korder)
 
         if ibatch == 0:
             # only on the first batch, we first get a new set of spikes from the residuals,
@@ -990,12 +1014,12 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
             # iW = int32(squeeze(iW))
 
             isort = cp.argsort(iW)  # sort by max abs channel
-            iW = iW[isort]
-            W = W[
+            iW = cp.asfortranarray(iW[isort])
+            W = cp.asfortranarray(W[
                 :, isort, :
-            ]  # user ordering to resort all the other template variables
-            dWU = dWU[:, :, isort]
-            nsp = nsp[isort]
+            ])  # user ordering to resort all the other template variables
+            dWU = cp.asfortranarray(dWU[:, :, isort])
+            nsp = cp.asfortranarray(nsp[isort])
 
         # decompose dWU by svd of time and space (via covariance matrix of 61 by 61 samples)
         # this uses a "warm start" by remembering the W from the previous iteration
@@ -1067,7 +1091,7 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
             WtW, iList = getMeWtW(W, U, Nnearest)
 
             # iW is the final channel assigned to each template
-            iW = cp.argmax(cp.abs(dWU[nt0min - 1, :, :]), axis=0)
+            iW = cp.asfortranarray(cp.argmax(cp.abs(dWU[nt0min - 1, :, :]), axis=0))
 
             # extract ALL features on the last pass
             Params[
@@ -1088,6 +1112,16 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
             for n in range(U.shape[1]):
                 # temporarily use U rather Urot until I have a chance to test it
                 ir.Wraw[:, :, n] = mu[n] * cp.dot(U[:, n, :], W[:, n, :].T)
+
+            if params.low_memory:
+                # Initialise array writers for the spike features of each cluster
+                feature_writers = [
+                    LargeArrayWriter(
+                        feature_path / f'spike_features_{i}',
+                        dtype=np.float32,
+                        shape=(NchanNear, Nrank, -1)
+                    ) for i in range(Nfilt)
+                ]
 
         if ibatch < niter - nBatches - 1:
             # during the main "learning" phase of fitting a model
@@ -1121,12 +1155,14 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
                     dWU0.shape,
                     order="F",
                 )
-                dWU = cp.concatenate((dWU, dWU0), axis=2)
+                dWU = cp.asfortranarray(cp.concatenate((dWU, dWU0), axis=2))
 
                 m = dWU0.shape[2]
                 # initialize temporal components of waveforms
-                W = _extend(
+                W = cp.asfortranarray(
+                    _extend(
                     W, Nfilt, Nfilt + m, W0[:, cp.ones(m, dtype=np.int32), :], axis=1
+                    )
                 )
 
                 # initialize the number of spikes with the minimum allowed
@@ -1148,15 +1184,8 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
         if ibatch > niter - nBatches - 1:
             # during the final extraction pass, this keeps track of all spikes and features
 
-            # we memorize the spatio-temporal decomposition of the waveforms at this batch
-            # this is currently only used in the GUI to provide an accurate reconstruction
-            # of the raw data at this time
-            ir.WA[..., k] = cp.asnumpy(W)
-            ir.UA[..., k] = cp.asnumpy(U)
-            ir.muA[..., k] = cp.asnumpy(mu)
-
             # we carefully assign the correct absolute times to spikes found in this batch
-            toff = nt0min + t0 + NT*k
+            toff = nt0min + t0 + NT*korder
             st = toff + st0
 
             st30 = np.c_[
@@ -1168,19 +1197,25 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
             ]
             # Check the number of spikes.
             assert st30.shape[0] == featW.shape[1] == featPC.shape[2]
+
+            # Sort by spike times
+            ix_sort = np.argsort(st30[:,0])
+            st30 = np.asfortranarray(st30[ix_sort])
+            featW = cp.asfortranarray(featW[:, ix_sort])
+            featPC = cp.asfortranarray(featPC[:, :, ix_sort])
+
             st3.append(st30)
             fW.append(featW)
             fWpc.append(featPC)
 
-            ntot = ntot + x0.size  # keeps track of total number of spikes so far
+            if params.low_memory:
+                # Save spike features individually for each cluster
+                for i in range(Nfilt):
+                    sub_array = cp.asfortranarray(featPC[:, :, id0 == i])
+                    feature_writers[i].append(sub_array)
 
-        if ibatch == niter - nBatches - 1:
-            # these next three store the low-d template decompositions
-            ir.WA = np.zeros((nt0, Nfilt, Nrank, nBatches), dtype=np.float32, order="F")
-            ir.UA = np.zeros(
-                (Nchan, Nfilt, Nrank, nBatches), dtype=np.float32, order="F"
-            )
-            ir.muA = np.zeros((Nfilt, nBatches), dtype=np.float32, order="F")
+
+            ntot = ntot + x0.size  # keeps track of total number of spikes so far
 
         if ibatch % 100 == 0:
             # this is some of the relevant diagnostic information to be printed during training
@@ -1208,6 +1243,11 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
     fW.close()
     fWpc.close()
 
+    if params.low_memory:
+        # Close the large array writers for each cluster's spike features
+        for writer in feature_writers:
+            writer.close()
+
     # just display the total number of spikes
     logger.info("Found %d spikes.", ntot)
 
@@ -1234,6 +1274,10 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
 
     # Number of spikes.
     assert ir.st3.shape[0] == fW.shape[-1] == fWpc.shape[-1]
+
+    # Save cluster times and IDs at this stage if requested
+    if params.save_temp_files:
+        np.save(ctx.context_path / 'temp_splits' / 'st3_learn.npy', ir.st3)
 
     # # this whole next block is just done to compress the compressed templates
     # # we separately svd the time components of each template, and the spatial components
@@ -1277,8 +1321,6 @@ def learnAndSolve8b(ctx, sanity_plots=False, plot_widgets=None, plot_pos=None):
         # cProjPC=ir.cProjPC,
         iNeigh=ir.iNeigh,
         iNeighPC=ir.iNeighPC,
-        WA=ir.WA,
-        UA=ir.UA,
         W=ir.W,
         U=ir.U,
         dWU=ir.dWU,