+++ pykilosort/params.py
@@ -59,13 +59,19 @@ class DatashiftParams(BaseModel):
 
 
 class KilosortParams(BaseModel):
+    
+    low_memory: bool = Field(
+        False, description='low memory setting for running chronic recordings'
+    )
+    
+    seed: t.Optional[int] = Field(42, description="seed for deterministic output")
+    
     preprocessing_function: str = Field('kilosort2', description='pre-processing function used choices'
                                                                  'are "kilosort2" or "destriping"')
+    
     save_drift_spike_detections: bool = Field(False, description='save detected spikes in drift correction')
+    
     perform_drift_registration: bool = Field(True, description='Estimate electrode drift and apply registration')
-    #TODO: Make algorithm deterministic for a fixed seed
-    seed: t.Optional[int] = Field(42, description="seed for deterministic output - \
-                                                   determinism coming soon")
 
     do_whitening: bool = Field(True, description='whether or not to whiten data, if disabled \
                                                  channels are individually z-scored')
@@ -102,12 +108,12 @@ class KilosortParams(BaseModel):
 
     sig_datashift: float = Field(20.0, description="sigma for the Gaussian process smoothing")
 
-    stablemode_enabled: bool = Field(False, description="make output more stable")
-    deterministicmode_enabled: bool = Field(False, description="make output deterministic by sorting spikes before applying kernels")
+    stable_mode: bool = Field(True, description="make output more stable")
+    deterministic_mode: bool = Field(True, description="make output deterministic by sorting spikes before applying kernels")
 
-    @validator("deterministicmode_enabled")
-    def validate_deterministicmode(cls, v, values):
-        if values.get("stablemode_enabled"):
+    @validator("deterministic_mode")
+    def validate_deterministic_mode(cls, v, values):
+        if values.get("stable_mode"):
             return v
         else:
             if v: